# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PGO Engine API",
    "description": "Secured with Basic Auth",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://pgo.otapp.live/api",
      "description": "Generated server url"
    }
  ],
  "security": [{ "basicAuth": [] }],
...
  "paths": {
...
    "/admin/v1/merchants/uid/{merchantUid}/api-keys": {
      "get": {
        "tags": ["Admin Merchant API Keys"],
        "summary": "List API keys",
        "description": "Retrieves all API keys belonging to the merchant",
        "operationId": "getMerchantApiKeys",
        "parameters": [
          {
            "name": "merchantUid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "200": {
            "description": "API keys retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseWrapperMerchantApiKeyResponseDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Admin Merchant API Keys"],
        "summary": "Create API key",
        "description": "Issues a new API key + secret for the merchant",
        "operationId": "createApiKey",
        "parameters": [
          {
            "name": "merchantUid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantApiKeyCreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapperMerchantApiKeyResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "format": "int32", "minimum": 0 },
          "size": { "type": "integer", "format": "int32", "minimum": 1 },
          "sort": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "MerchantApiKeyCreateRequestDto": {
        "type": "object",
        "description": "Optional parameters for issuing a merchant API key",
        "properties": {
          "validityDays": {
            "type": "integer",
            "format": "int32",
            "description": "Optional custom validity window in days (defaults to 365)",
            "example": 90,
            "maximum": 1095,
            "minimum": 1
          }
        }
      },
      "MerchantApiKeyResponseDto": {
        "type": "object",
        "description": "Contains API key credentials issued to a merchant",
        "properties": {
          "apiKey": {
            "type": "string",
            "description": "Public API key identifier (used in request headers)",
            "example": "pk_test_00000000000000000000000000"
          },
          "secretKey": {
            "type": "string",
            "description": "Secret part of the API key (returned only at creation time)",
            "example": "sk_test_00000000000000000000000000"
          },
          "expiresAt": {
            "type": "string",
            "description": "Expiration timestamp of the API key",
            "example": "2025-08-04T12:00:00"
          },
          "status": {
            "type": "string",
            "description": "Current status of the API key",
            "example": "ACTIVE"
          }
        }
      },
      "ResponseWrapperMerchantApiKeyResponseDto": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/MerchantApiKeyResponseDto" }
        }
      },
...
      "PageResponseWrapperMerchantApiKeyResponseDto": {
        "type": "object",
        "properties": {
          "status": { "type": "boolean" },
          "statusCode": { "type": "integer", "format": "int32" },
          "message": { "type": "string" },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MerchantApiKeyResponseDto"
            }
          },
          "pageNumber": { "type": "integer", "format": "int32" },
          "pageSize": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "last": { "type": "boolean" }
        }
      },
...
    },
    "securitySchemes": { "basicAuth": { "type": "http", "scheme": "basic" } }
  }
}